Convert Data to XTS Object

This function takes a data frame with a datetime column and one of two specified discharge columns, and converts it into an xts object, which is suitable for time series analysis. The datetime column is assumed to be named "dateTime" and should be in a format that can be converted to POSIXct. The discharge data can be in cubic feet per second ("Discharge_cfs") or cubic meters per second ("Discharge_cms").

convert_to_xts(data, column_choice = "Discharge_cfs")

Arguments

data

A data frame containing at least two columns: "dateTime" for the datetime of observations, and either "Discharge_cfs" or "Discharge_cms" for discharge data. The "dateTime" column is converted to POSIXct format if it is not already.

column_choice

A string specifying which column to use for the discharge data. Must be either "Discharge_cfs" or "Discharge_cms". Defaults to "Discharge_cfs".

Value

An xts object containing the selected discharge data indexed by the datetime of the observations.

See also

Examples

if (FALSE) {
  data <- data.frame(dateTime = Sys.time() + 0:4 * 3600,
                     Discharge_cfs = rnorm(5, mean = 100, sd = 10),
                     Discharge_cms = rnorm(5, mean = 2.83, sd = 0.28))
  xts_data <- convert_to_xts(data)
  print(xts_data)
}